Move some js to the head of the page
While it speeds up page load, putting bootstrap js lib at the bottom of
the doc created a weird loading effect where the content of the page
moved around.
Now the only js at the end are the libs used for the boids (for which
the delay poses no problem).