Minor modification of infobox handling
Change the approach for the boid infobox by putting the content in a
template at the end of the page and only instanciate it at the top of js
is enabled.

This allows, among other things, for a cleaner experience when using
weird web clients such as lynx or w3m.